requirements.txt
streamlit
python-docx
moviepy
requests
